Rename to GenericContainer
