+++ Libs/Optimize/itkParticleShapeMatrixAttribute.h
@@ -2,7 +2,7 @@
 
 #include "itkDataObject.h"
 #include "Observer.h"
-#include "itkParticleContainer.h"
+#include "GenericContainer.h"
 #include "itkParticleSystem.h"
 #include "itkWeakPointer.h"
 #include "vnl/vnl_matrix.h"
